requirements.txt
fastapi
uvicorn
python-dotenv
openai
chromadb
sentence-transformers
pandas
transformers
torch